+++ crates/swc_ecma_transforms/tests/fixture/legacy-only/regression-8041/input.ts
@@ -1,0 +1,4 @@
+export default class {
+    @foo
+    bar() { }
+}
