+++ crates/swc_ecma_transforms/tests/fixture/legacy-only/regression-8041/output.ts
@@ -1,0 +1,6 @@
+export default class _class {
+    bar() {}
+};
+__decorate([
+    foo
+], _class.prototype, "bar", null);
